extract function-decor-2 from class-decor
